Move auxil function to a new directory
