+++ tests/test_callbackquery.py
@@ -22,7 +22,11 @@ from datetime import datetime
 import pytest
 
 from telegram import Audio, Bot, CallbackQuery, Chat, Message, User
-from tests.conftest import check_defaults_handling, check_shortcut_call, check_shortcut_signature
+from tests.auxil.bot_method_checks import (
+    check_defaults_handling,
+    check_shortcut_call,
+    check_shortcut_signature,
+)
 
 
 @pytest.fixture(scope="function", params=["message", "inline"])
