+++ tests/test_photo.py
@@ -25,13 +25,12 @@ from telegram import Bot, InputFile, MessageEntity, PhotoSize, Sticker
 from telegram.error import BadRequest, TelegramError
 from telegram.helpers import escape_markdown
 from telegram.request import RequestData
-from tests.conftest import (
+from tests.auxil.bot_method_checks import (
     check_defaults_handling,
     check_shortcut_call,
     check_shortcut_signature,
-    data_file,
-    expect_bad_request,
 )
+from tests.conftest import data_file, expect_bad_request
 
 
 @pytest.fixture(scope="function")
